data-sample/requirements.txt
smart-open
libais
